requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
pydantic==2.4.2
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
alembic==1.12.0
pytest==7.4.2
httpx==0.25.0
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
redis==5.0.1
kafka-python==2.0.2
boto3==1.28.57
requests==2.31.0